requirements.txt
python-dotenv==1.1.0
openai==1.77.0
wget==3.2
pandas==2.2.3
transformers==4.51.3
opencv-python==4.11.0.86
Pillow==11.2.1
torch==2.7.0
torchvision==0.22.0
matplotlib==3.10.3
ipywidgets==8.1.7
qdrant-client==1.15.1
streamlit